Getting Started with REST and SDK/requirements.txt
azure-ai-textanalytics
azure-core
requests
